% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{extract_email_domain}
\alias{extract_email_domain}
\title{Extract email domain}
\usage{
extract_email_domain(email)
}
\arguments{
\item{email}{An email address as a string.}
}
\value{
A string with the domain part of an email address.
}
\description{
Extract email domain
}
\examples{
extract_email_domain("foo@mail.com") # mail.com
}
