% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_data.R
\docType{data}
\name{free_email_providers}
\alias{free_email_providers}
\title{List of free email providers}
\format{
A data frame with 119320 rows and 2 variables:
\describe{
\item{type}{type of provider: burner or non-burner}
\item{domain}{domain of the provider}
}
}
\source{
see data-raw
}
\usage{
free_email_providers
}
\description{
Dataset of free email providers.
}
\keyword{datasets}
