% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is_burner_email}
\alias{is_burner_email}
\title{Is email burner?}
\usage{
is_burner_email(email)
}
\arguments{
\item{email}{An email address as a string.}
}
\value{
A logical.
}
\description{
Is email burner?
}
\examples{
is_burner_email("foo@cuvox.de") # TRUE
is_burner_email("foo@gmail.com") # FALSE
}
