% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is_free_email}
\alias{is_free_email}
\title{Is email free?}
\usage{
is_free_email(email, include_burner = TRUE)
}
\arguments{
\item{email}{An email address as a string.}

\item{include_burner}{Should domains for burner emails be included. Default: TRUE}
}
\value{
A logical.
}
\description{
Is email free?
}
\examples{
is_free_email("foo@gmail.com") # TRUE
is_free_email("foo@myownservermeh.com") # FALSE
}
