% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is_valid_email}
\alias{is_valid_email}
\title{Is valid email?}
\usage{
is_valid_email(email)
}
\arguments{
\item{email}{An email address as a string.}
}
\value{
A logical.
}
\description{
Is valid email?
}
\examples{
is_valid_email("foo@mail.")
is_valid_email("foo@mail.com")
}
